+++ tests/test_text_onnx_embeddings.py
@@ -45,16 +45,16 @@ CANONICAL_VECTOR_VALUES = {
         [-0.0332, -0.0509, 0.0287, -0.0043, -0.0077]
     ),
     "jinaai/jina-embeddings-v2-base-de": np.array(
-      [-0.0085, 0.0417, 0.0342, 0.0309, -0.0149]
+        [-0.0085, 0.0417, 0.0342, 0.0309, -0.0149]
     ),
     "nomic-ai/nomic-embed-text-v1": np.array(
-        [0.0061, 0.0103, -0.0296, -0.0242, -0.0170]
+        [0.3708 ,  0.2031, -0.3406, -0.2114, -0.3230]
     ),
     "nomic-ai/nomic-embed-text-v1.5": np.array(
-        [-1.6531514e-02, 8.5380634e-05, -1.8171231e-01, -3.9333291e-03, 1.2763254e-02]
+        [-0.15407836, -0.03053198, -3.9138033, 0.1910364, 0.13224715]
     ),
     "nomic-ai/nomic-embed-text-v1.5-Q": np.array(
-        [-0.01554983, 0.0129992, -0.17909265, -0.01062993, 0.00512859]
+        [-0.12525563,  0.38030425, -3.961622 ,  0.04176439, -0.0758301]
     ),
     "thenlper/gte-large": np.array(
         [-0.01920587, 0.00113156, -0.00708992, -0.00632304, -0.04025577]
